requirements.txt
Quart==0.20.0
redis==5.2.1
aiohttp==3.11.14

fastapi>=0.68.0
uvicorn>=0.15.0
openai>=0.27.0
python-dotenv>=0.19.0
pydantic>=1.8.2
pyyaml>=6.0
dashscope==1.22.2